requirements.txt
networkx==3.1
matplotlib==3.7.1
folium==0.14.0
pandas==2.0.1
PyQt5==5.15.9
geopandas==0.13.2
shapely==2.0.1
pillow==10.0.0
